==> 03_Collect_annual_data.Rmd <==
---
title: "03 Collect data for time trend analysis"
output: 
  html_document:
    keep_md: true
    toc: true
    toc_depth: 3
    toc_float: true
    code_folding: hide
    df_print: paged   
---


## 0. Libraries
```{r}
library(tidyverse)
library(readxl)
library(broom)
# library(pander)
```

## 1. Folders and files    
Content of folders (see code)  
```{r}
dir("Datasett")
dir("Datasett/River data (from OKA)")
dir("Datasett/hydrografi")
dir("Datasett/Bløtbunn")
dir("Datasett/hardbunn_kopi")
dir("Datasett/Plankton")
```

## 2. River data   
Content of folders (see code)
```{r}
dir("Datasett/River data (from OKA)/Annual mean flows")
dir("Datasett/River data (from OKA)/Monthly loads")
dir("Datasett/River data (from OKA)/Concentrations (individual samples)")
dir("Datasett/River data (from OKA)/Monthly flow-weighted concentrations")
```

### a. Data of river loads
```{r}
df1 <- read_excel("Datasett/River data (from OKA)/Monthly loads/Storelva_monthly loads.xlsx")  

df2 <- read_excel("Datasett/River data (from OKA)/Monthly loads/Gjerstadelva_Nidelva_monthly loads.xlsx")  

df3 <- read_excel("Datasett/River data (from OKA)/Monthly loads/RIDx5_monthly loads.xlsx")  

# head(df1, 3)
# head(df2, 3)
# head(df3, 3)

colnames(df1) %>% dput()
colnames(df2) %>% dput()
colnames(df3) %>% dput()

df <- bind_rows(df1[-1,], df2[-1,], df3[-1,])

# colnames(df) %>% dput()
vars <- c("TrspTot TOTN", "TrspTot NO3-N", "TrspTot NH4-N", "TrspTot TOTP", 
          "TrspTot TOC", "TrspTot ALK", "TrspTot Ca", "DisTot")
for (var in vars)
  df[[var]] <- as.numeric(df[[var]])
df$Time <- with(df, lubridate::ymd(paste(Year, Month, "15")))

# Add "_" in column names (TrspTot Ca -> TrspTot_Ca)
colnames(df) <- sub(" ", "_", colnames(df), fixed = TRUE)

tb <- df %>% 
  gather("Variable", Value, TrspTot_TOTN:DisTot) %>%
  filter(!is.na(Value)) %>%
  xtabs(~Station_name + Variable, .)
tb

```

### b, Plot seasonal cycle
```{r}
gg <- df %>%
  gather("Variable", "Value",  TrspTot_TOTN:DisTot) %>%
  group_by(Station_name, Variable, Month) %>%
  summarise(Mean = mean(Value, na.rm = TRUE), Min = min(Value, na.rm = TRUE), Max = max(Value, na.rm = TRUE)) %>%
  ggplot(., aes(Month, Mean)) + 
    geom_ribbon(aes(ymin = Min, ymax = Max), fill = "lightgreen") +
    geom_line() + geom_point() +
    facet_grid(Variable~Station_name, scales = "free_y")
gg
```

### c. Quarterly means of all variables
```{r}
df2 <- df %>%
  mutate(Quarter = case_when(
    Month %in% 1:3 ~ 1,
    Month %in% 4:6 ~ 2,
    Month %in% 7:9 ~ 3,
    Month %in% 10:12 ~ 4
  )) %>%
  gather("Variable", "Value",  TrspTot_TOTN:DisTot) %>%
  group_by(Station_name, Variable, Year, Quarter) %>%
  summarise(Mean = mean(Value, na.rm = TRUE), Min = min(Value, na.rm = TRUE), Max = max(Value, na.rm = TRUE))

```


### d. Make quarterly site-specific variables
```{r}
df3 <- df2 %>%
  unite("VarQuarter", Variable, Quarter) %>%
  mutate(Station_short = substr(Station_name, 1, 4)) %>%
  unite("StVarQuarter", Station_short, VarQuarter) %>%
  arrange(StVarQuarter, Year)
```

### e. Test regression of every station / variable / quarter
```{r, fig.height = 18}
get_reg_coef <- function(df) {
  lm <- lm(Mean ~ Year, data = df)
  result <- subset(tidy(lm), term == "Year")
  # browser()
  result$Meanvalue <- mean(df$Mean)
  result$Perc_change_per_decade <- with(result, statistic/Meanvalue*10*100)
  result
  }

# Test
# get_reg_coef(df3 %>% filter(StVarQuarter %in% "Dram_DisTot_2" & !is.na(Mean)))


m_yeareffect <- df3 %>%
  filter(!is.na(Mean)) %>%
  group_by(StVarQuarter) %>%
  mutate(N_year = n()) %>%
  filter(N_year >= 10) %>%
  split(.$StVarQuarter) %>%
  map_df(~get_reg_coef(.), .id = "Variable")

ggplot(m_yeareffect, aes(Variable, statistic, fill = factor(p.value < 0.05))) +
         geom_col() +
  theme(axis.title.x = element_text(size = 4)) +
  coord_flip()

#   theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.2))
```

### f. Regression results, alternative plot
```{r, fig.width = 6, fig.height = 12}

# Helper functions
substr_right <- function(txt, n){              # n = no. of characters to extract from the right
  substr(txt, nchar(txt) - n + 1, nchar(txt))
}
substr_left <- function(txt, n_to_remove){     # n_to_remove = no. of characters to remove from the right
  substr(txt, 1, nchar(txt) - n_to_remove)
}

substr_mid <- function(txt, n_start, n_to_remove){ 
  substr(txt, n_start, nchar(txt) - n_to_remove)
}

# Sorts a character vector as the first 
sort_char_as_num <- function(txt){
  num <- stringr::str_extract(txt, "([^[[:blank:]]]+)")  %>% as.numeric()
  txt[order(num)]
}

# Test
# sort(c("10 %", "5 %", "20 %"))              # ordinary sort:    "10 %" "20 %" "5 %"
# sort_char_as_num(c("10 %", "5 %", "20 %"))  # sort_char_as_num: "5 %"  "10 %" "20 %"

factorlevels_from_num <- function(txt){
  factor(txt, levels = sort_char_as_num(unique(txt)))
  }

# Test
# substr_right("abcd", 2)
# substr_left(c("abcd", "abcdefgh"), 2)
# substr_mid(c("abcd", "abcdefgh"), 2, 2)

# quantile(m_yeareffect$Perc_change_per_decade)  %>% round(1)
# quantile(m_yeareffect$Perc_change_per_decade %>% abs())  %>% round(1)

# Make new data set as we need Quarter as a separate variable in the plot.
# Also we want to plot in the order of Variable_River, not River_Variable (keep each variable together in plot)
df <- m_yeareffect %>%
  mutate(Quarter = substr_right(Variable, 1), 
         River_Variable = substr_left(Variable, 2),
         Variable2 = substr_mid(Variable, 6, 2),
         Variable_River = paste0(Variable2, "_", substr(Variable, 1, 4))) %>%
  mutate(Slope_category = 
           case_when(statistic < 0 & p.value < 0.05 ~ "Decrease (P < 0.05)",
                     statistic < 0 & p.value > 0.05 ~ "Decrease (P > 0.05)",
                     statistic > 0 & p.value > 0.05 ~ "Increase (P > 0.05)",
                     statistic > 0 & p.value < 0.05 ~ "Increase (P < 0.05)")
  ) %>%
  mutate(Perc_change_category = 
           case_when(Perc_change_per_decade < 0.5 ~ "0 - 0.5 %",
                     Perc_change_per_decade >= 0.5 &  Perc_change_per_decade < 1 ~ "0.5 - 1 %",
                     Perc_change_per_decade >= 1 &  Perc_change_per_decade < 2 ~ "1 - 2 %",
                     Perc_change_per_decade >= 2 &  Perc_change_per_decade < 10 ~ "2 - 10 %",
                     Perc_change_per_decade >= 10 &  Perc_change_per_decade < 20 ~ "10 - 20 %") 
  )  %>%
  mutate(Perc_change_category = factorlevels_from_num(Perc_change_category))


xtabs(~Perc_change_category, df)

gg1 <- ggplot(df, aes(Quarter, Variable_River, size = abs(Perc_change_per_decade), color = Slope_category)) +
  geom_point() +
  scale_size_continuous("% change per decade", breaks = c(0, 1, 2, 5, 10, 20)) +
  scale_color_manual("Statistical significance", values = c("darkblue", "lightblue", "pink", "red3"))
# gg1
  
gg2 <- ggplot(df, aes(Quarter, Variable_River, size = Perc_change_category, color = Slope_category)) +
  geom_point() +
  scale_size_manual("% change per decade", values = 3:8) +
  scale_color_manual("Statistical significance", values = c("darkblue", "lightblue", "pink", "red3"))
gg2
```
