==> 06_Timeseries_regression_Theil-Sen.Rmd <==
---
title: "06 - Time series regressions using Mann-Kendall and Theil-Sen"
output:
  html_document:
    keep_md: true
    toc: true
    toc_depth: 3
    toc_float: true
    code_folding: hide
    df_print: paged   
---

Non-parametric tests for trends over time  
* Performed both on
    + Annual data (ending in _a, plus rivergroup_springflood)
    + Seasonal data (for seasons separately)
* For seasonal data, note that hydrography + plankton + river seasons start in February in order to get a "full spring bloom" in quarter 1 from February to April
* Uses data from script 05b_ (example: df_rivers_summ_a)
* produces summarized data from different datatypes onn annual (Dat_a) and quarterly (Dat_q) for use in regression models (08_regression)
* produces plots of time series trends (Kap 3.2) for quarterly and annual data for the report
* Test = Mann-Kendall trend test (for P) and Theil-Sen's slope estimator (for percent change per year)   


DHJ: you borrowed much of this code from   'file:///C:/Data/Referanseelver_2018/Climate/05_Analysis_gridded_data_MK_TheilSen.html'    



## 0. Libraries
```{r}
library(tidyverse)
library(readxl)
library(broom)
library(lubridate)
# install.packages("rkt")
library(rkt)        # rkt() - computes the Mann-Kendall tests and Theil-Sen's slope estimator

#citation ("rkt")
#citation ("ggplot2")

# library(pander)
save_plots <- TRUE

```

## 1. Data (annual and partly seasonal)  
### a. Read data
* rivers, hydrography and plankton are gathered from annual datasets produced in 05b_
* ON HOLD: indices and ordinations for plankton, soft and hard-bottom. 
```{r}

# rivers
df_rivers_summ_a <- read.csv("Data_produced/05_df_rivers_summ_a.csv")
#df_rivergroup_springflood <- read.csv("Data_produced/05_df_rivergroup_springflood.csv")

#hydrography
df_hydro_summ_a <- read.csv("Data_produced/05_df_hydro_summ_a.csv")

# plankton groups
df_plank_summ_a <- read.csv("Data_produced/05_df_plank_summ_a.csv")

# Hard-bottom lower growth limit - EQR from MSMDI (lower growth limit index)
df_lgm <- readr::read_csv2("Datasett/Hardbunn_KOPI/HBanalysesett2.csv")

# Soft bottom indices and ordinations
df_blot_ind <- read.csv("Data_produced/01_df_blot_ind.csv")
# rename soft-bottom stations
levels(df_blot_ind$STAS)[levels(df_blot_ind$STAS)=="B05"] <- "BR1"
levels(df_blot_ind$STAS)[levels(df_blot_ind$STAS)=="B35"] <- "BT44"


# NOT USE: importing the ordinations for plotting in Mann_kendall 
#Plankton ordination (Bloom2): 
# df_plank_ord <- readr::read_csv2("Data_produced/P02ord.csv")
# # retrieving softbottom ordinations
# df_blot_ord_05 <- readr::read_csv2("Data_produced/BBordSites_05.csv")
# #names(df_blot_ord_05)
# df_blot_ord_35 <- readr::read_csv2("Data_produced/BBordSites_35.csv")
# #names(df_blot_ord_35)

# # hard-bottom indices and ordinations
# df_hard <- readr::read_csv2("Datasett/Hardbunn_KOPI/HBanalysesett.csv")
# # rename hard-bottom stations
# df_hard$Site[which(df_hard$Site==407)]<-"HT113"
# df_hard$Site[which(df_hard$Site==410)]<-"HR104"
# #str(df_hard)
# # select only the two stations we use
# df_hard <- df_hard %>%
#   filter(Site %in% c("HT113", "HR104"))


```

### b. Collect data  
* River data
    + discharge divided by 1 million, the rest divided by 1000
```{r}

# df1 = rivers
str(df_rivers_summ_a)
df1 <- df_rivers_summ_a %>%
  mutate(Discharge = Discharge/1E3) %>%
  gather("Variable", "Value", TotN:Si) %>%
  mutate(Variable = paste0("River", "_", Variable)) %>%
  mutate(Value = Value/1E3) %>%
  dplyr::select(Variable, Year, Value)

# HFR: removed these lines because only use distant river group and not include spring flood
  #mutate(River_type = paste0("River_", River_type),
   #      Variable = sub("TrspTot_", "", Variable, fixed = TRUE)) %>%
  #mutate(Variable = paste0(River_type, "_", Variable),
         #Value = Value/1E3) %>%

# head(df1a)
#df1 %>% group_by(Variable) %>% summarise(mean(Value, na.rm = TRUE))

#df1b <- df_rivergroup_springflood %>%
 #rename(Floodmax = DisTot_max_rel, FloodmaxMonth = DisTot_max_month) %>%
# gather("Variable", "Value", Floodmax, FloodmaxMonth) %>%
 # mutate(River_type = paste0("River_", River_type)) %>%
#  mutate(Variable = paste0(River_type, "_", Variable)) %>%
 # select(Variable, Year, Value)
# head(df1b)

#df2= hydrography
#str(df_hydro_summ_a)
df2 <- df_hydro_summ_a %>%
  gather("Variable", "Value", Temperature, Salinity, O2, PO4, Si, Chla:DIN) %>%
  mutate(Variable = paste0("Hydro", "_", Variable, "_", Depth)) %>%
  filter(!(Variable %in% c("Hydro_Secchi_Intermediate", "Hydro_Secchi_Deep"))) %>%
  droplevels() %>%
  dplyr::select(Variable, Year, Value)

# df3a= plankton groups
#str (df_plank_summ_a)
df3a <- df_plank_summ_a %>%
  gather("Variable", "Value", Diatoms:Flagellates) %>%
  mutate(Variable = paste0("Plankton", "_", Variable)) %>%
  dplyr::select(Variable, Year,Value)

df4 <- df_lgm %>%
  filter(Site %in% c(407, 410)) %>%
  mutate(Station = case_when(Site == 407 ~ "HT113",
                             Site == 410 ~ "HR104")) %>%
  mutate(Variable = paste0("HardBottom_Lowergrowth_", Station)) %>%
  rename(Value = EQR) %>%
  dplyr::select(Variable, Year, Value)


#df5 = soft-bottom NQI
str (df_blot_ind)
df5 <- df_blot_ind %>%
   gather("Variable", "Value", NQI1) %>%
   mutate(Variable = paste0("SoftBottom", "_", Variable, "_", STAS)) %>%
   select(Variable, Year, Value)

#NOT USE:  Ordinations
# # df3b= plankton DCA1
# df3b <- df_plank_ord %>%
#   gather("Variable", "Value", DCA1) %>%
#   mutate(Variable = paste0("Plankton", "_", Variable)) %>%
#   select(Variable, Year,Value)

# # df4b + c = Soft-bottom ordinations for stations BR1 and BT44
# df4b <- df_blot_ord_05 %>%
#   gather("Variable", "Value", DCA1) %>%
#   mutate(Variable = paste0("SoftBottom", "_", Variable, "_", "BR1")) %>%
#   select(Variable, Year,Value)
# 
# df4c <- df_blot_ord_35 %>%
#   gather("Variable", "Value", DCA1) %>%
#   mutate(Variable = paste0("SoftBottom", "_", Variable, "_", "BT44")) %>%
#   select(Variable, Year, Value)
# 
# # df5 =  Hard bottom DCA1 and LGD for different species (can we summarize species?)
# #str (df_hard)
# df5 <- df_hard %>%
#   gather("Variable", "Value", DCA1, "LGD.Halidrys siliquosa":"LGD.Rhodomela confervoides") %>%
#   mutate(Variable = paste0("HardBottom", "_", Variable)) %>%
#   select(Variable, Year,Value)

# collect all datasets

#df3b, df4a, df4b, df4c, df5 #ON HOLD
dat_a <- rbind(df1, df2, df3a, df4, df5) %>% as.data.frame()
head(dat_a)


```

### c. Read quarterly data
```{r}
df_rivers_summ_q <- read.csv("Data_produced/05_df_rivers_summ_q.csv")
df_hydro_summ_q <- read.csv("Data_produced/05_df_hydro_summ_q.csv")
df_plank_summ_q <- read.csv("Data_produced/05_df_plank_summ_q.csv")

#str (df_rivers_summ_q)
#str (df_hydro_summ_q)
#str (df_plank_summ_q)

```

### d. Collect quarterly data  
* River data
    + discharge divided by 1 million, the rest divided by 1000
```{r}

df1 <- df_rivers_summ_q %>%
  mutate(Discharge = Discharge/1E3) %>%
  gather("Variable", "Value", TotN:Si) %>%
  mutate(Variable = paste0("River", "_", Variable)) %>%
  mutate(Value = Value/1E3) %>%
  select(Variable, Year, Quarter, Value)

head(df1)
df1 %>% group_by(Variable) %>% summarise(mean(Value, na.rm = TRUE))

#str(df_hydro_summ_q)
df2 <- df_hydro_summ_q %>%
  gather("Variable", "Value", Temperature, Salinity, O2, PO4, Si, Chla:DIN) %>%
  mutate(Variable = paste0(Variable, "_", Depth)) %>%
  filter(!(Variable %in% c("Secchi_Intermediate", "Secchi_Deep"))) %>%
  select(Variable, Year, Quarter, Value) %>%
  select(Variable, Year, Quarter, Value)

#head(df2)

df3 <- df_plank_summ_q %>%
  gather("Variable", "Value", Diatoms:Flagellates) %>%
  select(Variable, Year, Quarter, Value)
head(df3)

dat_q <- rbind(df1, df2, df3) %>% as.data.frame()
```

e. Save collected data
```{r}
write.csv(dat_a, "Data_produced/06_dat_a.csv", row.names = FALSE, quote = FALSE)
write.csv(dat_q, "Data_produced/06_dat_q.csv", row.names = FALSE, quote = FALSE)

```

## 3a. Define trend analysis functions
One for annual data, one for seasonal
```{r}

trend_analysis_a <- function(variable, data = dat_a){
  df <- data %>%
    filter(Variable %in% variable & !is.na(Value))
  result <- rkt(df$Year, df$Value)
  data.frame(Variable = variable, Quarter = "Annual", 
             P = result$sl, Estimate = result$B, 
             Change_perc = 100*result$B/quantile(df$Value, 0.5), stringsAsFactors = FALSE)
}

#debugonce(trend_analysis_a)
#trend_analysis_a("River_Distant_TOTN")

trend_analysis_q <- function(variable, quarter, data = dat_q){
  df <- data %>%
    filter(Variable %in% variable & Quarter %in% quarter & !is.na(Value))
  result <- rkt(df$Year, df$Value)
  data.frame(Variable = variable, Quarter = as.character(quarter), 
             P = result$sl, Estimate = result$B, 
             Change_perc = 100*result$B/quantile(df$Value, 0.5), stringsAsFactors = FALSE)
}

#trend_analysis_q("River_Distant_TOTN", 2)

```

## 3b. Annual data, perform analysis  
```{r}
# ?map

df <- dat_a %>%
  filter(!is.na(Value)) %>%
  group_by(Variable) %>%
  summarise(N = n()) %>%
  filter(N >= 8)
 
df_result_list <-  1:nrow(df) %>% map(~trend_analysis_a(df$Variable[.], data = dat_a))

df_result_a <- bind_rows(df_result_list)


```


## 3c. Quarterly data, perform analysis  
```{r}
# ?map
df <- dat_q %>%
  filter(!is.na(Value)) %>%
  group_by(Variable, Quarter) %>%
  summarise(N = n()) %>%
  filter(N >= 8)
 df_result_list <- 
  1:nrow(df) %>% map(~trend_analysis_q(df$Variable[.], df$Quarter[.], data = dat_q))
 df_result_q <- bind_rows(df_result_list)

```


## 4. Show results
### a. Most extreme changes
```{r}

# head = mest negative = st�rst reduksjon over tid.
df_result_q %>% arrange(Change_perc) %>% tail()
df_result_q %>% arrange(Change_perc) %>% head()

```

## 5. Save results
```{r}

write.csv(df_result_a, "Data_produced/06_df_result_a.csv",
          row.names = FALSE, quote = FALSE)
write.csv(df_result_q, "Data_produced/06_df_result_q.csv",
          row.names = FALSE, quote = FALSE)
```

## 6. Some plots
### a. Dot plot theme
```{r}
# Code and theme from http://www.joyce-robbins.com/blog/2016/06/02/datavis-with-rdrawing-a-cleveland-dot-plot-with-ggplot2/
theme_dotplot <- theme_bw(14) +
    theme(axis.text.y = element_text(size = rel(.75)),
    	axis.ticks.y = element_blank(),
        axis.title.x = element_text(size = rel(.75)),
        panel.grid.major.x = element_blank(),
        panel.grid.major.y = element_line(size = 0.5),
        panel.grid.minor.x = element_blank())
```

### b. Plot one season
```{r, fig.height=9}
qrt <- 2        
df <- df_result_q %>% filter(Quarter == qrt) 

ggplot(df, aes(x = Change_perc, y = reorder(Variable, Change_perc), size = factor(P < 0.05))) +
	geom_point(color = "blue") +
	scale_x_continuous(limits = c(-12, 4), breaks = seq(-12, 4, 2)) +
  scale_size_manual(values = c(2,4)) +
	# scale_size_continuous(limits = c(-12, 4), breaks = seq(-12, 4, 2)) +
	theme_dotplot +
	xlab("\nChange per year (%)") +
	ylab("Variable\n") +
	ggtitle(paste("Changes in quarter", qrt))
```


### c. Plot for all seasons
* HFR: rather color code variables by datatype (river, hydro, plankton), instead of having it as prefix to variable?
```{r, fig.width = 7, fig.height=9}
df_mean <- df_result_q  %>%
  group_by(Variable) %>%
  summarise(Mean_change = mean(Change_perc)) %>%
  arrange(Mean_change)

df <- df_result_q
df$Variable <- factor(df$Variable, levels = df_mean$Variable)
df$Quarter <- factor(df$Quarter)
df$P <- factor(df$P < 0.05, labels = c("Not significant", "Significant"))
# range(df$Change_perc)

cols <- RColorBrewer::brewer.pal(4, "RdBu")[c(3,4,1,2)]

# create the plot
ggplot(df, aes(x = Change_perc, y = Variable, fill = Quarter, size = P)) +
  geom_vline(xintercept = 0) +
	geom_point(pch = 21) +
  # scale_color_brewer(palette = "RdBu", direction = -1) +
  scale_fill_manual(values = cols) +
  scale_size_manual("Trend test", values = c(2.5,4)) +
	scale_x_continuous(limits = c(-11, 7), breaks = seq(-10, 6, 2)) +
	theme_dotplot +
	xlab("\nChange per year (%)") +
	ylab("Variable\n")

if (save_plots)
  ggsave ("Figures_rapp/MannKendall_q.png", width = 7, height = 8, dpi=500)


```

### d. Plot trends for annual data
```{r, fig.width = 7, fig.height=9}
# create the plot
df <- df_result_a
df$P <- factor(df$P < 0.05, labels = c("Not significant", "Significant"))
df <- df[order(df$Change_perc),]
# range(df$Change_perc)

gg <- ggplot(df, aes(x = Change_perc, y = reorder(Variable, Change_perc), fill = P, size = P)) +
  geom_vline(xintercept = 0) +
	geom_point(pch = 21) +
  scale_fill_manual("Trend test", values = c("lightblue", "firebrick")) +
  scale_size_manual("Trend test", values = c(2.5,4)) +
	scale_x_continuous(limits = c(-11, 4), breaks = seq(-10, 4, 2)) +
	theme_dotplot +
	xlab("\nChange per year (%)") +
	ylab("Variable\n")

gg + theme(
    legend.position = c(.35, .98),
    legend.justification = c("right", "top"),
    legend.box.just = "left",
    legend.box.background = element_rect(),
    legend.margin = margin(6, 6, 6, 6),
    axis.title.x = element_text(size = rel(1)),
    axis.text.y = element_text(color = df$P),
    axis.text.y = element_text(face = c("plain","bold")[df$P]),
    axis.title.y = element_blank()
  )

if (save_plots)
  ggsave ("Figures_rapp/MannKendall_a.png", width = 7, height = 9, dpi=500)


```

### e. Plots of annual data, showing ONLY significant trends (for ASLO presentation)
```{r}

# create the plot
df <- df_result_a
df$P <- factor(df$P < 0.05, labels = c("Not significant", "Significant"))
df <- df[order(df$Change_perc),]
# range(df$Change_perc)

# select only variables with significant trends, and make new ID column for data type
df_s <- df %>%
  filter(P %in% c("Significant")) %>%
  mutate(ID=sapply(strsplit(Variable, split='_', fixed=TRUE),function(x) (x[1]))) %>%
  mutate(ID = recode(ID, 'Plankton' = 'Bio', 'HardBottom' = 'Bio', "SoftBottom" = "Bio"))

# getting nice colors for the data types
library(colorspace)
df_s$ID <- as.factor (df_s$ID) 
df_s$color <- factor(df_s$ID,
levels=c("River", "Bio", "Hydro"),
labels=rainbow_hcl(3))


gg <- ggplot(df_s, aes(x = Change_perc, y = reorder(Variable, Change_perc))) +
  geom_vline(xintercept = 0) +
	geom_point(pch = 19, size = 3) +
  #scale_fill_manual("Trend test", values = c("lightblue", "firebrick")) +
  #scale_size_manual("Trend test", values = c(2.5, 4)) +
	scale_x_continuous(limits = c(-11, 4), breaks = seq(-10, 4, 2)) +
	theme_dotplot +
	xlab("\nChange per year (%)") +
	ylab("Variable\n")

gg

gg + theme(
    legend.position = c(.35, .98),
    legend.justification = c("right", "top"),
    legend.box.just = "left",
    legend.box.background = element_rect(),
    legend.margin = margin(6, 6, 6, 6),
    axis.title.x = element_text(size = rel(1)),
    axis.text.y = element_text(colour = as.character (df_s$color), face = "bold"),
    axis.title.y = element_blank()
  )


if (save_plots)
  ggsave ("Figures_presentation/MannKendall_a.png", width = 7, height = 9, dpi=500)

```

## 7. Make table of trends for annual data 
```{r}
#install.packages("kableExtra")

library(kableExtra)

tab <- df_result_a %>%
  mutate(Var_order = sub("River", "01", Variable),             # Ordering River first, then Hydro etc.
         Var_order = sub("Hydro", "02", Var_order),
         Var_order = sub("Plankton", "03", Var_order),
         Var_order = sub("HardBottom", "04", Var_order),
          Var_order = sub("SoftBottom", "05", Var_order)) %>%
  arrange(Var_order) %>%
    mutate(
    car = row.names(.),
    Estimate = cell_spec(sprintf("%.4g", Estimate), bold = ifelse(P < 0.05, TRUE, FALSE)),
    Change_perc = cell_spec(sprintf("%.2f", Change_perc, 2), bold = ifelse(P < 0.05, TRUE, FALSE)),
    P = cell_spec(ifelse(P > 0.001, sprintf("%.2g", P), "<0.001"))
  ) %>%
  select(Variable, Estimate, Change_perc, P) %>%
  kable(escape = F) %>%
  kable_styling("condensed", full_width = F) 


if (save_plots)
  writeLines(tab, "Figures_rapp/MannKendall_a_table.html")
# result is opened in web browser, Ctrl-A (to select all) and then copy-pasted to Word/Googledocs

tab

```
