==> 09_Map_stations_rivers.Rmd <==
---
title: "09 Map with stations and rivers"
author: "Helene"
date: "13 november 2018"
output: html_document
---

## 0. Libraries
```{r}
library(ggplot2)

#install.packages("ggrepel")
library(ggrepel)

# packages hvis teste eksempel med googlemaps
#library(RgoogleMaps)
#library(ggmap)
#library(grid)

# packages for old script
#library (maps)
#install.packages("mapdata")
#library(mapdata)
```

### 1. Make dataframe with stations (lat, long) 
```{r}

# making dataframe (positions in WGS84)
df2 <- data.frame("Name" = c("Arendal","Grimstad","Arendal","Tromøya","Prestholmen",
                             "Glomma","Drammenselva","Numedalslågen","Skienselva","Gjerstadelva",
                             "Storelva","Nidelva"), 
                 "Type" = c("Hydro/Plankton","Soft-bottom","Soft-bottom","Hard-bottom","Hard-bottom",
                            "River", "River", "River", "River", "River", "River", "River"),
                 "ID" = c("VT5","BR1", "BT44","HT113", "HR104",
                          "GLO","DRA","NUM","SKI","GJE","STO","NID" ), 
                 "Lon" = c(8.8330, 8.6295, 9.0312, 8.9445, 8.5372, 
                           11.134, 10.007, 10.070, 9.608, 9.078, 8.854, 8.637), 
                 "Lat" = c(58.3870, 58.3253, 58.4038, 58.5132, 58.2732,
                           59.279, 59.752, 59.101, 59.204, 58.763, 58.632, 58.407))

str(df2)
print (df2)



```

### 1. Plotte kart
* Dag: jeg prøvde å merke de ID (kode for ulike stasjone) i plottet, for eks med geom_label eller med shape (men blir for mange), men får det ikke helt til. det må jo med inne i aes() på et vis, men fikk ikke helt til koplingen der... 
* tenkte også jeg skulle prøve med googlemap, slik at vi kan få et større område med elver osv. men kan være det blir for bråkete ja... 
```{r}

# kystverket data, høyoppløslig kyslinje, men ikke sve/dk
load("K:/Avdeling/214-Oseanografi/DHJ/Data/Maps/Norway_coastline_longlat2.RData")

# test med Kystverkets kart

# 
ggplot(data=df2, aes(x = Lon, y = Lat, color = as.character (Type))) + 
  annotation_map(norway_coast_longlat2, aes(long, lat), fill = "darkolivegreen3")  +
  coord_map("lambert", parameters = c(58.7, 9.3), xlim = c(8, 11.5), ylim = c(58, 60)) +
  geom_point() +
  labs(x = "Longitude", y= "Latitude", color = "Station type")

# example with ggrepel
ggplot(data=df2, aes(x = Lon, y = Lat, color = as.character (Type))) + 
  annotation_map(norway_coast_longlat2, aes(long, lat), fill = "darkolivegreen3")  +
  coord_map("lambert", parameters = c(58.7, 9.3), xlim = c(8, 11.5), ylim = c(58, 60)) +
  geom_point() +
  geom_label_repel(data = df2,
            aes(Lon, Lat,label = ID,group = ID),
            color = 'black',
            size  = 3,
            box.padding = 0.5, point.padding = 0.5) +
  labs(x = "Longitude", y= "Latitude", color = "Station type")

ggsave ("Figures_rapp/Map_prelim.png", width = 8, height = 6, dpi=500)


# openstreetmap, større område
load("K:/Avdeling/214-Oseanografi/DHJ/Data/Maps/coastline_skagerrak_coords_longlat.RData")
plot(y~x, type = "l", coastline_skagerrak_coords_longlat, 
     xlim = c(5, 14), ylim = c(53, 60))



# eksempel med googlemap fra Hordaklim (ikke tilpasset): http://yarkerconsulting.com/index.php/blog/15-google-maps-and-r
library(RgoogleMaps)
library(ggplot2)
library(ggmap)
require(grid)
# Another way of plotting

basemap <- get_map(location=c(lon=mean(lon),lat=mean(lat)), zoom = 8, maptype='roadmap', source='google',crop=TRUE)

ggmap(basemap)

map1 <- ggmap(basemap, extent='panel', base_layer=ggplot(data=stdata, aes(x=lon, y=lat)))

# map showing point size based on altitude

map.alt <- map1+geom_point(aes(size=stdata$alt),color="darkblue")+scale_size(range=c(2,9),breaks=c(10,100,500,1000))

# add plot labels

map.alt <- map.alt + labs(x ="Longitude", y="Latitude", size = "Altitude")

# add title theme

map.alt <- map.alt + theme(plot.title = element_text(hjust = 0, vjust = 1, face = c("bold")))+theme(legend.position="bottom")

print(map.alt) 


```

## OLD: code used to make map in 2013 article.
```{r, eval = FALSE}

# draw map
map("worldHires", xlim = c (4, 15), ylim = c(53,60), fill = TRUE, col = "grey", ylab = "Latitude", xlab = "Longitude")
# add axes
map.axes()
mtext("Latitude",side=2, line=3)
mtext("Longitude",side=1, line=3)
# scale bar, uten ratio
#map.scale(ratio = FALSE)

#Arendal St. 2 (58Â° 23Â´N 08Â° 49Â´E)
points(y=58.23,x=08.49,col=1,pch=16,cex=1.5)
text ("Arendal", y=58.4,x=8)

# add country names
text ("Norway", y=59.5,x=8, cex=1.2)
text ("Sweden", y=58.6,x=13, cex=1.2)
text ("Denmark", y=56.2,x=9.1, cex=1.1)
text ("U.K.", y=52.5,x=1, cex=1.2)
text ("Germany", y=53.4,x=10, cex=1.2)

# oceans
text ("Skagerrak", y=57.9,x=9.5)
text ("Kattegat", y=56.7,x=11.6)
text ("North Sea", y=56,x=5.5)



# Kristiandsand 58Â°10â€²12â€³N 8Â°0â€²22â€³E
#decimal  58.17, 8.006111
points(y=58.10,x=08,col=1,pch=18, cex = 1.5)
#text ("K", y=58.1,x=7)

# Hanstholm 57Â°Â 7â€²Â 12â€³Â N, 8Â°Â 37â€²Â 12â€³Â E
# decimal 57.12, 8.62
points(y=57.12,x=08.62,col=1,pch=18, cex = 1.5)
#text ("H", y=57,x=8.5)

# source water stations
points(y=54.18,x=07.89,col=1,pch=15, cex = 1.5)
text ("Helgoland", y=54.2,x=7.9)

points(y=56.4,x=12.07,col=1,pch=15, cex = 1.5)
text ("Anholt", y=56.5,x=12.2)

points(y=57.59,x=9.96,col=1,pch=15, cex = 1.5)
text ("Hirtshals", y=57.5,x=9.9)
```
