==> 10_Regression_plankton.Rmd <==
---
title: "10 Regression plankton"
output: 
  html_document:
    keep_md: true
    toc: true
    toc_depth: 3
    toc_float: true
    code_folding: hide
    df_print: paged   
---

# Regression of plankton

Regression analysis of plankton variables  
  
Parts 0 + 1a-e (libraries + data) is identical to script 08 (copied 26.11 kl. 22:00)  


## 0. Libraries
```{r}
# to check which libraries are set
#.libPaths()
#.libPaths("C:/Data/R/R-3.5.1/library")
library(tidyverse)
library(readxl)
library(MASS)
library (ggplot2)
library(visreg)
# library(vegan)
# library(lme4) # for mixed effect models
#install.packages("standardize")
library(standardize) # standardize data for modelling
#install.packages("afex")
# library(afex) # adding p-values to lmer output
#install.packages("sjPlot") 
library(sjPlot) # effectplots lmer
# library(pander)
#install.packages("glmulti")
# library(glmulti)
#install.packages("rJava") # NB! I needed to installe Java for Windows x64 here: https://www.oracle.com/technetwork/java/javase/downloads/jdk8-downloads-2133151.html

# install.packages("MuMIn") 
library(MuMIn)
#library(rJava)
```

## 1. Data  
### a. Read
```{r}
dat_a <- read.csv("Data_produced/06_dat_a.csv")
dat_q <- read.csv("Data_produced/06_dat_q.csv")
```

### b. Check variables
```{r}
unique(dat_a$Variable)
```

### c. Make data on broad format
```{r}
dat <- 
  dat_a %>% 
  spread(Variable, Value)
```


### d. Add NAO data
```{r}
df_nao <- read.table("https://climatedataguide.ucar.edu/sites/default/files/nao_station_djfm.txt", skip = 1, header = FALSE)
colnames(df_nao) <- c("Year", "Winter_NAO")

dat <- left_join(dat, df_nao)

```

### e. Wind data ??? Dag you mentioned you had files for Færder and Torungen?


### f. Checking missing years in dataset
* NB! TSM has been interpolated for 2012-2013, we decided not to interpolate the _Deep nutrients and POM, because these are at the end of the time series
```{r}

str(dat)

check <-  dat %>%
  summarise_all(funs(sum(is.na(.))))

complete.cases(dat)
# there are missing values for four first years (1990-1994) and the last four years (2013-2016)
  #River TOC missing in 1990 and 1991, 
  #P measurements from hydro missing 1990
  #Plankton group data start in 1994
  #River_Si starts in 1995
  # DIN_deep, POC_deep, PON,_deep and POP_deep is missing from 2013-2016

# Suggestion: 
  #1. We cut 1990 and 1991 from the dataset (many missing variables)
  #2. we cut variables that have more than one year missing (have them separate below so we can test)

# NB! keep in mind that the deep POC and PON are the variables with highest upward trends in the MannKendal test. but a possible proxy could be the TSM_deep which is sampled over the whole time period, and has fairly high correlation with POC (0.63). High correlations also between the deep and intermediate values for POC (0.75), PON (0.74) and DIN (0.72). 
plot (dat$Hydro_TSM_Deep, dat$Hydro_POC_Deep)
plot (dat$Hydro_TSM_Deep, dat$Hydro_PON_Deep)
plot (dat$Hydro_POC_Deep, dat$Hydro_POC_Intermediate)
plot (dat$Hydro_DIN_Deep, dat$Hydro_DIN_Intermediate)
cor(dat$Hydro_POC_Surface, dat$Hydro_POC_Intermediate, use = "complete.obs")
#cor(dat$Hydro_TSM_Deep, dat$Hydro_POC_Deep, use = "complete.obs")
#cor(dat$Hydro_PON_Deep, dat$Hydro_PON_Intermediate, use = "complete.obs")
#cor(dat$Hydro_DIN_Deep, dat$Hydro_DIN_Intermediate, use = "complete.obs")

# filter the datasets according to suggestion above
cols <- c("Hydro_DIN_Deep", "Hydro_POC_Deep", "Hydro_PON_Deep", "Hydro_POP_Deep", "River_Si")

dat_sel <- dat %>% 
  filter(Year != 1990 & Year != 1991) %>% 
  # dplyr::select(-starts_with('Plank')) %>% 
  dplyr::select(-one_of(cols))    

# Years with no missing cases - 1992-1993 is missing, but that is because of the plankton variables
rbind(dat_sel$Year, complete.cases(dat_sel))
tab <- apply(is.na(dat_sel), 2, sum)
tab[tab > 0]

```


# Plankton data
* excluding Hydro_Chla_Surface for all and two models for each group (a w/ secchi included, b w/ secchi excluded)
```{r}
dat_sel2 <- dat_sel[complete.cases(dat_sel),]

options(na.action = "na.fail")

save_plots <- FALSE

#NB! exclude Chla_surface for all

```

## a. Diatoms
* include Si_Surface
```{r}


### 1a Diatoms
full_model <- lm(Plankton_Diatoms ~ River_TOC + River_SPM + Winter_NAO + 
                    Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface
                 + Hydro_PO4_Surface + Hydro_Si_Surface + Hydro_TotP_Surface +
                   Hydro_TotN_Surface + Hydro_Temperature_Surface, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, rank = "BIC") 
model_diatoms <- get.models(min_model, 1)[[1]]
summary (model_diatoms)
# Plot model
# model diagnostics
par(mfrow = c(2,2), mar = c(4,5,2,1))
plot(model_diatoms)
# plot_model(model_diatoms, type = "diag")
# plot model estimates (neutral line; vertical intercept that indicates no effect)
# plot_model(model_diatoms, sort.est = TRUE)

#effektplot
# plot_model(model_diatoms, type = "slope", show.data = TRUE)
par(mar = c(4,6,2,1))
visreg(model_diatoms,  points = list(cex = 1), ylab = "")
mtext("Diatom abundance", 2, line = 4)

if(save_plots){
  png("Figures_rapp/Regressions_Plankton_diatoms.png", width = 18, height = 15, unit = "cm", res = 400)
  par(mfrow = c(2,2), mar = c(4,5,2,1), oma = c(0,3,0,0))
  visreg(model_diatoms,  points = list(cex = 1), ylab = "")
  mtext("Diatom abundance", 2, line = 1, outer = TRUE)
  dev.off()
}

tab_model (model_diatoms)
# , file = "Figures_rapp/Regressions_POC_Surf_table.html"
```

## b. Dinoflagellates
### Including Hydro_Secchi_Surface (but not Chl a)
```{r}
# 1b

full_model <- lm(Plankton_Dinoflagellates ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface +
                   Hydro_Secchi_Surface, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, rank = "BIC") 
model_dinoflag01 <- get.models(min_model, 1)[[1]]
summary (model_dinoflag01)
# Plot model
# model diagnostics
par(mfrow = c(2,2), mar = c(4,5,2,1))
plot(model_dinoflag01)
# plot_model(model_dinoflag01, type = "diag")
# plot model estimates (neutral line; vertical intercept that indicates no effect)
# plot_model(model_dinoflag01, sort.est = TRUE)

#effektplot
# plot_model(model_dinoflag01, type = "slope", show.data = TRUE)
par(mfrow = c(2,3), mar = c(4,5,2,1))
visreg(model_dinoflag01,  points = list(cex = 1))

if(save_plots){
  png("Figures_rapp/Regressions_Plankton_dinoflag01.png", width = 25, height = 15, unit = "cm", res = 400)
  par(mfrow = c(2,3), mar = c(4,5,2,1), oma = c(0,3,0,0))
  visreg(model_dinoflag01,  points = list(cex = 1), ylab = "")
  mtext("Dinoflagellate abundance", 2, line = 1, outer = TRUE)
  dev.off()
}

tab_model (model_dinoflag01)


```

### Excluding Hydro_Chla_Surface + Hydro_Secchi_Surface
```{r}
# 1b
full_model <- lm(Plankton_Dinoflagellates ~ River_TOC + River_SPM + Winter_NAO + 
                    Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface + 
                    Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface +
                    Hydro_Temperature_Surface, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, rank = "BIC") 
model_dinoflag02 <- get.models(min_model, 1)[[1]]
summary (model_dinoflag02)
# Plot model
# model diagnostics
par(mfrow = c(2,2), mar = c(4,5,2,1))
plot(model_dinoflag02)
# plot_model(model_dinoflag02, type = "diag")
# plot model estimates (neutral line; vertical intercept that indicates no effect)
# plot_model(model_dinoflag02, sort.est = TRUE)

#effektplot
# plot_model(model_dinoflag02, type = "slope", show.data = TRUE)
par(mfrow = c(2,3), mar = c(4,5,2,1))
visreg(model_dinoflag02,  points = list(cex = 1))


if(save_plots){
  png("Figures_rapp/Regressions_Plankton_dinoflag02.png", width = 25, height = 7.5, unit = "cm", res = 400)
  par(mfrow = c(1,3), mar = c(4,5,2,1), oma = c(0,3,0,0))
  visreg(model_dinoflag02,  points = list(cex = 1), ylab = "")
  mtext("Dinoflagellate abundance", 2, line = 1, outer = TRUE)
  dev.off()
}


tab_model (model_dinoflag02)


```



## c. Flagellates
### Including Hydro_Secchi_Surface
```{r}
# 1b
full_model <- lm(Plankton_Flagellates ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface +
                   Hydro_Secchi_Surface, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, rank = "BIC") 
model_flag01 <- get.models(min_model, 1)[[1]]
summary (model_flag01)
# Plot model
# model diagnostics
par(mfrow = c(2,2), mar = c(4,5,2,1))
plot(model_flag01)
# plot_model(model_flag01, type = "diag")
# plot model estimates (neutral line; vertical intercept that indicates no effect)
plot_model(model_flag01, sort.est = TRUE)

#effektplot
# plot_model(model_flag01, type = "slope", show.data = TRUE)
par(mfrow = c(2,3), mar = c(4,5,2,1))
visreg(model_flag01,  points = list(cex = 1))



if(save_plots){
  png("Figures_rapp/Regressions_Plankton_flag01.png", width = 18, height = 15, unit = "cm", res = 400)
  par(mfrow = c(2,2), mar = c(4,5,2,1), oma = c(0,3,0,0))
  visreg(model_flag01,  points = list(cex = 1), ylab = "")
  mtext("Flagellate abundance", 2, line = 1, outer = TRUE)
  dev.off()
}


tab_model (model_flag01)


```

### Excluding Hydro_Chla_Surface + Hydro_Secchi_Surface
```{r}
# 1b
full_model <- lm(Plankton_Flagellates ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, rank = "BIC") 
model_flag02 <- get.models(min_model, 1)[[1]]
summary (model_flag02)
# Plot model
# model diagnostics
par(mfrow = c(2,2), mar = c(4,5,2,1))
plot(model_flag02)
# plot_model(model_flag02, type = "diag")
# plot model estimates (neutral line; vertical intercept that indicates no effect)
plot_model(model_flag02, sort.est = TRUE)

#effektplot
# plot_model(model_flag02, type = "slope", show.data = TRUE)
par(mfrow = c(2,3), mar = c(4,5,2,1))
visreg(model_flag02,  points = list(cex = 1))


if(save_plots){
  png("Figures_rapp/Regressions_Plankton_flag02.png", width = 18, height = 15, unit = "cm", res = 400)
  par(mfrow = c(2,2), mar = c(4,5,2,1), oma = c(0,3,0,0))
  visreg(model_flag02,  points = list(cex = 1), ylab = "")
  mtext("Flagellate abundance", 2, line = 1, outer = TRUE)
  dev.off()
}

tab_model (model_flag02)


```


## d. Plankton including year effect (effectively like detrending data)
```{r}
#
### Diatoms
#
full_model <- lm(Plankton_Diatoms ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface +
                   Hydro_Secchi_Surface +
                   Year, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, fixed = "Year", rank = "BIC") 

# All models with dBIC < 2

subset(min_model, delta <= 2)

# Best model
get.models(min_model, 1)[[1]] %>% summary()

#
### Plankton_Dinoflagellates
#
full_model <- lm(Plankton_Dinoflagellates ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface +
                   Hydro_Secchi_Surface +
                   Year, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, fixed = "Year", rank = "BIC") 

# All models with dBIC < 2
subset(min_model, delta <= 2)

# Best model
get.models(min_model, 1)[[1]] %>% summary()



#
### Plankton_Flagellates
#
full_model <- lm(Plankton_Flagellates ~ River_TOC + River_SPM + Winter_NAO +
                   Hydro_Salinity_Surface + Hydro_POC_Surface + Hydro_DIN_Surface +
                   Hydro_PO4_Surface + Hydro_TotP_Surface + Hydro_TotN_Surface + 
                   Hydro_Temperature_Surface +
                   Hydro_Secchi_Surface +
                   Year, 
                  data = dat_sel2) 
# summary(full_model)
# finne min model
min_model <- dredge (full_model, fixed = "Year", rank = "BIC") 

# All models with dBIC < 2
subset(min_model, delta <= 2)

# Best model
get.models(min_model, 1)[[1]] %>% summary()



```

## e. Interpretation of dinoflagellate model
### I. Plot
```{r}
lev <- c("Hydro_DIN_Surface", "Hydro_Secchi_Surface", "Hydro_Temperature_Surface", 
           "Hydro_TotP_Surface", "River_SPM", "River_TOC", "Plankton_Dinoflagellates")
dat_sel2 %>% 
  dplyr::select(Year, Hydro_DIN_Surface, Hydro_Secchi_Surface, Hydro_Temperature_Surface, 
         Hydro_TotP_Surface, River_SPM, River_TOC, Plankton_Dinoflagellates) %>%
  gather(Variable, Value, -Year) %>%
  mutate(Variable = factor(Variable, levels = lev)) %>%
  ggplot(aes(Year, Value)) + geom_smooth() + geom_point() +
  facet_wrap(~Variable, scales = "free_y")
```

### II. Effect of each variable on slope of dinoflagellates vs. year   
Set one variable to mean value, and print  1 - (new slope)/ (original slope) 
```{r}
dat_pred <- list(
  dat_sel2 %>% mutate(Hydro_DIN_Surface = mean(Hydro_DIN_Surface)),
  dat_sel2 %>% mutate(Hydro_Secchi_Surface = mean(Hydro_Secchi_Surface)),
  dat_sel2 %>% mutate(Hydro_Temperature_Surface = mean(Hydro_Temperature_Surface)),
  dat_sel2 %>% mutate(Hydro_TotP_Surface = mean(Hydro_TotP_Surface)),
  dat_sel2 %>% mutate(River_SPM = mean(River_SPM)),
  dat_sel2 %>% mutate(River_TOC = mean(River_TOC))
)

vars <- c("Hydro_DIN_Surface", "Hydro_Secchi_Surface", "Hydro_Temperature_Surface", 
          "Hydro_TotP_Surface", "River_SPM", "River_TOC")

cat("Fraction of dinoflagellate decline caused by said variable:\n")
for (i in 1:6){
  yr <- 1994:2016
  pred_orig <- predict(model_dinoflag01)
  pred_man <- predict(model_dinoflag01, new = dat_pred[[i]])
  #plot(yr, pred_orig)
  #points(yr, pred_man, pch = 19, col = 2)
  slope_orig <- coef(lm(pred_orig ~ yr))[2]
  slope_man <- coef(lm(pred_man ~ yr))[2]
  cat(vars[i], ":", 1 - slope_man/slope_orig, "\n")
}

i <- 4
  yr <- 1994:2016
  pred_orig <- predict(model_dinoflag01)
  pred_man <- predict(model_dinoflag01, new = dat_pred[[i]])
  plot(yr, pred_orig, type = "b")
  points(yr, pred_man, pch = 19, col = 2, type = "b")
  points(yr, dat_sel2$Plankton_Dinoflagellates, pch = 19, col = 4)

```
